app/requirements.txt
flask==3.0.3
boto3==1.34.162
redis==5.0.7
mysql-connector-python==9.0.0
python-json-logger==2.0.7
gunicorn==22.0.0

